Website/requirements.txt
Flask>=1.0,<=1.1.2
networkx==2.5
matplotlib
python-constraint